VideoFrame: split hpp/cpp
